zisk-repl
zik-repl is a Read–Eval–Print Loop (REPL) built in Python that allows users to interact directly with the Zisk language in real time. Its main purpose is to facilitate language exploration for students and neurodiverse users, enabling them to experiment with logical structures and programming concepts in a friendly, accessible, and highly customizable environment.